+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file is used to list changes made in each version of cookbook-openstack-obj
 * Upgrading to Juno
 * Upgrading berkshelf from 2.0.18 to 3.1.5
 * Bump Chef gem to 11.16
+* Add swift.conf template
 
 ## 9.0.3
 * Bugfix run_command exitstatus